+++ example/src/main/java/org/nervos/ckb/transaction/NumberUtils.java
@@ -13,12 +13,4 @@ public class NumberUtils {
     }
     return integers;
   }
-
-  public static String getZeros(int length) {
-    StringBuilder sb = new StringBuilder();
-    for (int i = 0; i < length; i++) {
-      sb.append("0");
-    }
-    return sb.toString();
-  }
 }